refactor(upstash): follow OS Protocol folder convention
Move cache implementation to system/data/cache/ to mirror
the SDK's protocol structure, making it self-documenting
which interface is being implemented.
