+++ extensions/upstash/src/index.d.ts
@@ -10,4 +10,4 @@ export {
   type Cache,
   type CacheEntry,
   type CacheStats,
-} from './cache'
+} from './system/data/cache'
