+++ extensions/upstash/src/index.ts
@@ -5,4 +5,4 @@
  * Provides distributed caching with Set-based invalidation.
  */
 
-export * from './cache'
+export * from './system/data/cache'
