+++ extensions/upstash/src/system/data/cache/index.d.ts
