+++ extensions/upstash/tsup.config.ts
@@ -5,7 +5,7 @@ import { defineConfig } from 'tsup'
 export default defineConfig({
   entry: {
     index: 'src/index.ts',
-    cache: 'src/cache/index.ts',
+    'system/data/cache': 'src/system/data/cache/index.ts',
   },
   format: ['cjs', 'esm'],
   // DTS disabled - manual type declarations provided in src/*.d.ts
